requirements.txt
streamlit
google-generativeai
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv
langdetect       # for basic language detection (optional)